refactor: Board와 JanggiGame의 책임을 분리
