+++ src/main/java/janggi/domain/board/Board.java
@@ -1,7 +1,6 @@
 package janggi.domain.board;
 
 import janggi.domain.piece.Piece;
-import janggi.domain.piece.Team;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -9,7 +8,6 @@ import java.util.Map;
 
 public class Board {
     private final Map<Position, Piece> board;
-    private Turn turn = new Turn(Team.HAN);
 
     public Board(Map<Position, Piece> board) {
         this.board = board;
@@ -20,16 +18,17 @@ public class Board {
     }
 
     public void move(Position from, Position to) {
-        validatePieceExistsAt(from);
         Piece piece = board.get(from);
-        validateCurrentTurn(piece);
 
         if (!piece.canMove(from, to, this)) {
             throw new IllegalArgumentException("해당 기물은 이동할 수 없습니다.");
         }
 
         movePiece(from, to);
-        changeTurn();
+    }
+
+    public Piece findPiece(Position position) {
+        return board.get(position);
     }
 
     public boolean hasPieceAt(Position position) {
@@ -47,28 +46,8 @@ public class Board {
         return positionPieces;
     }
 
-    private boolean isCurrentTeamPiece(Piece piece) {
-        return turn.isCurrentTeam(piece.getTeam());
-    }
-
     private void movePiece(Position from, Position to) {
         board.put(to, board.get(from));
         board.remove(from);
-    }
-
-    private void changeTurn() {
-        this.turn = turn.changeTurn();
-    }
-
-    private void validatePieceExistsAt(Position from) {
-        if (!hasPieceAt(from)) {
-            throw new IllegalArgumentException("해당 출발 위치에는 기물이 존재하지 않습니다");
-        }
-    }
-
-    private void validateCurrentTurn(Piece piece) {
-        if (!isCurrentTeamPiece(piece)) {
-            throw new IllegalArgumentException("해당 기물은 현재 턴의 진영 기물이 아닙니다.");
-        }
     }
 }
